glibc: Move `sys/statvfs.h` definitions to `src/new`
This follows the ongoing migration into `src/new`, which reduces
duplication and models glibc's source tree.

The `__f_unused` padding field is only declared on some targets, so it
is gated behind a `cfg`.

`riscv32` and `riscv64` declared `__f_spare` as pub, which was
inconsistent with the other declarations and had no grounded reason to
stay, so it is now private.

Links:

* https://github.com/sailfishos-mirror/glibc/blob/04e750e75b73957cf1c791535a3f4319534a52fc/sysdeps/unix/sysv/linux/bits/statvfs.h
* https://github.com/sailfishos-mirror/glibc/blob/04e750e75b73957cf1c791535a3f4319534a52fc/io/sys/statvfs.h

(backport <https://github.com/rust-lang/libc/pull/5434>)
(cherry picked from commit fc8a345915b2594fbc1e5c8d1cf28648f3bbe852)
