+++ src/new/mod.rs
@@ -201,6 +201,8 @@ cfg_if! {
         pub use net::route::*;
         #[cfg(target_env = "gnu")]
         pub use signal::*;
+        #[cfg(target_env = "gnu")]
+        pub use sys::statvfs::*;
     } else if #[cfg(target_vendor = "apple")] {
         #[cfg(target_os = "macos")]
         pub use net::bpf::*;
